refactor(util): extract file URL parser
