+++ src/providers/google/base.ts
@@ -19,7 +19,7 @@ import path from 'path';
 import cliState from '../../cliState';
 import { importModule } from '../../esm';
 import logger from '../../logger';
-import { parseFileUrl } from '../../util/functions/loadFunction';
+import { parseFileUrl } from '../../util/functions/parseFileUrl';
 import { maybeLoadToolsFromExternalFile } from '../../util/index';
 import { getNunjucksEngine } from '../../util/templates';
 import { MCPClient } from '../mcp/client';
