+++ src/util/file.ts
@@ -11,7 +11,7 @@ import { importModule } from '../esm';
 import logger from '../logger';
 import { runPython } from '../python/pythonUtils';
 import { isJavascriptFile } from './fileExtensions';
-import { parseFileUrl } from './functions/loadFunction';
+import { parseFileUrl } from './functions/parseFileUrl';
 import { safeResolve } from './pathUtils';
 import { renderVarsInObject } from './render';
 import { loadYaml } from './yamlLoad';
