+++ src/util/functions/parseFileUrl.ts
@@ -1,0 +1,48 @@
+import { isJavascriptFile } from '../fileExtensions';
+
+// Matches the leading slash + Windows drive prefix from canonical `file:///C:/...`
+// URLs (e.g. `/C:/` or `/C:\`). Only stripped on Windows so POSIX paths that
+// legitimately start with `/X:` (a directory literally named `X:`) are preserved.
+const WIN32_DRIVE_PREFIX = /^\/[A-Za-z]:[\\/]/;
+
+function normalizeFilePath(filePath: string): string {
+  if (process.platform === 'win32' && WIN32_DRIVE_PREFIX.test(filePath)) {
+    return filePath.slice(1);
+  }
+  return filePath;
+}
+
+/**
+ * Extracts the file path and function name from a file:// URL
+ * @param fileUrl The file:// URL (e.g., "file://path/to/file.js:functionName")
+ * @returns The file path and optional function name
+ */
+export function parseFileUrl(fileUrl: string): { filePath: string; functionName?: string } {
+  if (!fileUrl.startsWith('file://')) {
+    throw new Error('URL must start with file://');
+  }
+
+  const urlWithoutProtocol = fileUrl.slice('file://'.length);
+  const lastColonIndex = urlWithoutProtocol.lastIndexOf(':');
+
+  if (lastColonIndex > 1) {
+    const candidateFilePath = urlWithoutProtocol.slice(0, lastColonIndex);
+
+    // Only executable function files support a :functionName suffix. This preserves
+    // colons that are part of a valid file or directory name on POSIX systems.
+    if (!isJavascriptFile(candidateFilePath) && !candidateFilePath.endsWith('.py')) {
+      return {
+        filePath: normalizeFilePath(urlWithoutProtocol),
+      };
+    }
+
+    return {
+      filePath: normalizeFilePath(candidateFilePath),
+      functionName: urlWithoutProtocol.slice(lastColonIndex + 1),
+    };
+  }
+
+  return {
+    filePath: normalizeFilePath(urlWithoutProtocol),
+  };
+}
